% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{as.df}
\alias{as.df}
\title{Convert list of lists into data.frame}
\usage{
as.df(x)
}
\arguments{
\item{x}{the list of lists}
}
\value{
a \code{data.frame}
}
\description{
The list of lists should only contain lists with the same element names.
}
\examples{
x <- as.df(list(list(a=1,b="foo"),list(a=2,b="bar")))
}
