% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{c2q}
\alias{c2q}
\title{Convert from Raw coordinate to Quadrant address}
\usage{
c2q(x)
}
\arguments{
\item{x}{a raw coordinate (e.g. B15) (do not directly use on vectors!)}
}
\value{
the quadrant plate coordinate
}
\description{
Converts address tags for 384-well plates from the 
raw coordinate system (e.g. B15) to the quadrant system (e.g. C_A08).
}
\examples{
c2q("B15")
}
