% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{canRead}
\alias{canRead}
\title{Can read file}
\usage{
canRead(filename)
}
\arguments{
\item{filename}{the name of the file}
}
\value{
boolean
}
\description{
Determines whether the given file exists and is readable
}
\examples{
\dontrun{
canRead("foobar.txt")
}
}
