% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{colAlpha}
\alias{colAlpha}
\title{Add alpha channel}
\usage{
colAlpha(color, alpha)
}
\arguments{
\item{color}{a predefined color string (e.g. "firebrick")}

\item{alpha}{a number between 0 and 1 for the alpha channel value}
}
\description{
Adds an alpha channel (i.e. transparency) to a predefined color
}
\examples{
transparentChartreuse <- colAlpha("chartreuse3",0.3)
}
