% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{combo}
\alias{combo}
\title{Set of subsets}
\usage{
combo(l)
}
\arguments{
\item{l}{a list or vector}
}
\value{
a list of lists containing all possible subsets of the input
}
\description{
Generates the set of all possible subsets for a given list or vector
}
\examples{
combo(1:4)
}
