% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{drawPvalBracket}
\alias{drawPvalBracket}
\title{Helper function for drawing p-values over barplots}
\usage{
drawPvalBracket(p, i, j, h = 1.1, s = 0.02)
}
\arguments{
\item{p}{p-value}

\item{i}{x-coordinate of the left bar}

\item{j}{x-coordinate of the right bar}

\item{h}{height (y-coordinate) at which to draw the bracket}

\item{s}{spacer size (distance between brackets; size of feet)}
}
\value{
nothing
}
\description{
Helper function for drawing p-values over barplots
}
