% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{extract.groups}
\alias{extract.groups}
\title{Extract regex groups (local)}
\usage{
extract.groups(x, re)
}
\arguments{
\item{x}{A vector of strings from which to extract the groups.}

\item{re}{The regular expression defining the groups}
}
\value{
A \code{matrix} containing the group contents, 
     with one row for each element of x and one column for each group.
}
\description{
Locally excise regular expression groups from string vectors.
I.e. only extract the first occurrence of each group within each string.
}
\keyword{expression}
\keyword{groups}
\keyword{regular}
