% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{getArg}
\alias{getArg}
\title{Get CLI argument}
\usage{
getArg(name, default = NULL, required = FALSE)
}
\arguments{
\item{name}{The name of the command line argument}

\item{default}{If no value was given by the user, default to this}

\item{required}{Whether the argument is required or not. If TRUE,
an error is raised when no value was provided.}
}
\description{
Retrieves a user-supplied argument command-line argument
with a given name. Argument syntax: name=value
}
\examples{
\dontrun{
infile <- getArg("inputFile",required=TRUE)
userIQ <- getArg("userIQ",default=0)
}
}
