% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{\%>\%}
\alias{\%>\%}
\title{envelop a function in another}
\usage{
f \%>\% g
}
\arguments{
\item{f}{the inner funtion}

\item{g}{the outer function}

\item{a}{new function that simply wraps the inner inside the outer function}
}
\description{
This is intended to help with apply statements, where one may want 
to pass a function that simply wraps one function in another
}
\examples{
#an example matrix with some NA values in it
mat <- sample(c(NA,1:5),100,replace=TRUE)|>matrix(nrow=20)
#find all rows that don't have NA values in them
cleanRows <- apply(mat, 1, is.na \%>\% any \%>\% `!`)
}
