% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{is.color}
\alias{is.color}
\title{Checks if a string is a valid color}
\usage{
is.color(str)
}
\arguments{
\item{str}{a character vector}
}
\value{
a logical vector of the same length indicating which elements are valid colors
}
\description{
Given a character vector, this function will check whether each element in the vector
is a valid color descriptor, such as "black", "chartreuse2", or "#ff0000"
}
\examples{
is.color(c(NA, "black", "blackk", "1", "#00", "#000000"))
}
\references{
\url{http://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation/13290832#13290832}
}
