% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{ith.rank}
\alias{ith.rank}
\title{Get i'th rank from list}
\usage{
ith.rank(values, i, high = TRUE)
}
\arguments{
\item{values}{a numerical vector}

\item{i}{the rank}

\item{high}{whether to rank by highest or lowest values.}
}
\value{
the ith ranked value
}
\description{
Retieve the i'th ranked item from a numerical vector
}
\examples{
vals <- rnorm(100,0,1)
ith.rank(vals,4)
}
