% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{mcc}
\alias{mcc}
\title{Matthew's correlation coefficient (MCC)}
\usage{
mcc(t, scores, truth)
}
\arguments{
\item{t}{the score threshold}

\item{scores}{vector of scores for each measured item}

\item{truth}{\code{logical} vector classifying each item as a 
member of the hidden true or false classes}
}
\value{
a vector listing the MCC value, the precision, and the recall
}
\description{
Calculate Matthew's correlation coeffient (MCC). 
See \url{https://en.wikipedia.org/wiki/Matthews_correlation_coefficient}
}
\examples{
patientHasDisease <- sample(c(TRUE,FALSE),100,replace=TRUE)
patientDiganosticScore <- sapply(patientHasDisease,
   function(d) if (d) rnorm(1,20,3) else rnorm(1,18,3)
)
mccval <- mcc(21,patientDiganosticScore,patientHasDisease)
}
