% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{new.cluster.map}
\alias{new.cluster.map}
\title{Cluster mapper}
\usage{
new.cluster.map(n)
}
\arguments{
\item{n}{The number of elements to cluster.}
}
\value{
the mapper object
}
\description{
Constructor for an object supporting simple connected-component-clustering
}
\details{
The process starts with n objects, each in their own cluster. Whenever a link
is between two objects is reported, their clusters are merged.
Contains the following functions:
\begin{itemize}
  \item \code{addLink(i,j)}: Creates a new link between items i and j. Whenever a link
     is created, the clusters encompassing the two objects are merged.
  \item \code{getClusters()}: Returns a list of lists representing the clusters
  \item \code{getIdxOf(i)}: Returns the cluster index of a given object.
\end{itemize}
}
\examples{
cmap <- new.cluster.map(10)
cmap$addLink(1,5)
cmap$addLink(3,5)
cmap$addLink(1,5)
cmap$getClusters()
}
