% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{q2c}
\alias{q2c}
\title{Convert from Quadrant to Coordinate adress}
\usage{
q2c(x)
}
\arguments{
\item{x}{a quadrant coordinate (e.g. C_A08) (do not directly use on vectors!)}
}
\value{
the raw plate coordinate
}
\description{
Converts address tags for 384-well plates from the quadrant system (e.g. C_A08)
to the raw coordinate system (e.g. B15).
}
\examples{
q2c("C_A08")
}
