% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{rowApply}
\alias{rowApply}
\title{apply a function to a data.frame's rows}
\usage{
rowApply(tbl, f)
}
\arguments{
\item{tbl}{a data.frame}

\item{f}{a function with parameters matching the column names}
}
\value{
a list of the return values of f
}
\description{
This is intended to work similar to apply(x,1,f), but instead of 
coercing the row into vector, they are provided as individual function arguments.
This helps preserve the original data types
}
