% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{runningFunction}
\alias{runningFunction}
\title{Generic function for running means/medians etc across 2D data}
\usage{
runningFunction(x, y, nbins, fun = mean, logScale = FALSE)
}
\arguments{
\item{x}{the axis over which bins will be formed}

\item{y}{the dimension on which the function 'fun' is applied}

\item{nbins}{The number of bins to use}

\item{fun}{the function that will be applied to the y values in each bin}

\item{logScale}{whether or not bin-size will be at log scale across x}

\item{width}{the width of the bins}
}
\value{
a matrix with two columns: bin centroid, and value of fun
}
\description{
Generic function for running means/medians etc across 2D data
}
