% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{to.df}
\alias{to.df}
\title{Convert row-bound lists to data.frame}
\usage{
to.df(x)
}
\arguments{
\item{x}{the result of the rbind call.}
}
\value{
a \code{data.frame}
}
\description{
Running \code{rbind} on lists with the same element names
yields a datastructure very similar to a \code{data.frame}, but does not
provide the same full functionality. This function converts such 
objects to a real dataframe.
}
\examples{
x <- rbind(list(a=1,b="foo"),list(a=2,b="bar"))
y <- to.df(x)
}
