% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{toChars}
\alias{toChars}
\title{Deconstruct a string into single characters}
\usage{
toChars(str)
}
\arguments{
\item{str}{the input string}
}
\value{
a vector of characters
}
\description{
Converts a single character string into a a vector of individual characters
}
\examples{
nucleotides <- toChars("ACTTGCTAAACTTGA")
}
