% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{topoScatter}
\alias{topoScatter}
\title{Topographical scatterplot}
\usage{
topoScatter(
  x,
  y,
  resolution = 20,
  thresh = 5,
  topoCol = colorRampPalette(c("black", "red", "yellow"))(20),
  maxFreq = NULL,
  xlim = range(x, na.rm = TRUE),
  ylim = range(y, na.rm = TRUE),
  log = "",
  ...
)
}
\arguments{
\item{x}{a vector of x-coordinates}

\item{y}{a vector of y-coordinates (must match x in length)}

\item{resolution}{the resolution of the plot, that is into how many bins each axis will
be sub-divided. Defaults to 20.}

\item{thresh}{a threshold of how many data points must fall into a bin before it they 
replaced with a density-indicating color}

\item{topoCol}{a vector of colors defining a color gradient to use for the density map}

\item{xlim}{the x-axis range (see plot()).}

\item{ylim}{the y-axis range (see plot()).}

\item{log}{As in plot(). "x" draws the x-axis in log scale, "y" draws the y-axis in log-scale, 
"xy" draws both in log scale, "" draws both linearly.}
}
\description{
Draw a 'topographical scatterplot' or 'scatter-heatmap'
}
\examples{
x <- rnorm(10000,0,1)+10
y <- rnorm(10000,0,1)+10
topoScatter(x,y,resolution=60,xlab="foo",ylab="bar",thresh=2)
topoScatter(x,y,log="xy",resolution=60,xlab="foo",ylab="bar")

}
