% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{weightedAverage}
\alias{weightedAverage}
\title{join multiple datapoints weighted by stdev}
\usage{
weightedAverage(ms, sds, dfs, ws = (1/sds)/sum(1/sds))
}
\arguments{
\item{ms}{data means}

\item{sds}{data stdevs}

\item{dfs}{degrees of freedom for each data point}

\item{ws}{datapoint weights. By default these get auto-calculated from sds}
}
\value{
a vector containing the joint mean (mj), joint stdev (sj), and combined degrees of freedom
}
\description{
join multiple datapoints weighted by stdev
}
