% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{zbind}
\alias{zbind}
\title{3D-bind matrices}
\usage{
zbind(...)
}
\arguments{
\item{...}{Any number of matrices of the same size}
}
\value{
A 3D array of the bound matrices
}
\description{
Binds matrices of same size together to a 3D array, analogously
to cbind and rbind.
}
