% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{errorBars}
\alias{errorBars}
\title{Helper function for drawing error bars on barplots}
\usage{
errorBars(xs, val, err, l = 0.01, ...)
}
\arguments{
\item{xs}{vector of x coordinates of bars}

\item{val}{the values (heights) of the bars}

\item{err}{the size amount of error associated with each bar}

\item{l}{the length of the terminators of the bar}

\item{...}{any other graphical parameters}
}
\description{
Helper function for drawing error bars on barplots
}
