% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yogitools.R
\name{errorBars}
\alias{errorBars}
\title{Helper function for drawing error bars}
\usage{
errorBars(xs, val, err, l = 0.01, vertical = TRUE, topToBottom = TRUE, ...)
}
\arguments{
\item{xs}{vector of x coordinates of data points (In case of vertical=FALSE, 
these are interpreted as y-coordinates)}

\item{val}{the values (heights) of the data points.}

\item{err}{the size amount of error associated with each bar}

\item{l}{the length of the terminators of the bar}

\item{vertical}{whether the bars are vertical or horizontal}

\item{topToBottom}{whether the error is represented as the distance
between the top and bottom of the bar, or the middle and the top.}

\item{...}{any other graphical parameters}
}
\description{
Helper function for drawing error bars
}
